[drape] Store 8 bool flags in bits; store zoom and rank as uint8_t.

+++ drape/overlay_tree.cpp
@@ -144,7 +144,7 @@ void OverlayTree::InvalidateOnNextFrame()
   m_frameCounter = kInvalidFrame;
 }
 
-void OverlayTree::StartOverlayPlacing(ScreenBase const & screen, int zoomLevel)
+void OverlayTree::StartOverlayPlacing(ScreenBase const & screen, uint8_t zoomLevel)
 {
   ASSERT(IsNeedUpdate(), ());
   TBase::Clear();
